requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-bootstrap==3.3.7.1
python-dotenv==1.0.0
pytest==7.4.3
python-docx==1.1.2
